move types into node-saml
